+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## [Not Released v0.9]
 
+## 2022-01-27/avz
+### changed
+- added navigation bar to the header component
+
 ## 2022-01-26/ml
 ### changed
 - kolkov/editor updated to 2.0.0 (Ivy Version)